Ammend tests
Callbacks are immediately invoked if the player is not running.
This causes the sessions and graveyard to differ from before but prevents ghosting in production.
We could add a passthrough for the test, but that would be anti to testing doctrine. There shouldn't be made special considerations for tests.

For commit 07dc909b39af7865a61df84b1b5a049534c865ab